REPO-1 fix ambiguous reference

+++ repository/workflow/src/test/java/org/onehippo/repository/scxml/AbstractActionTest.java
@@ -158,7 +158,7 @@ public class AbstractActionTest {
                 // or #goImmutable() method.
             }
 
-            assertEquals(55, eval(getParameter("oneExpression")));
+            assertEquals(55, (long) eval(getParameter("oneExpression")));
         }
     }
 }
